Add ounce-force/ozf force unit
This unit is extremely common in motor specifications in US when
specifing motor toque using the ambiguous form of `oz in`, it should
be `ozf in` to follow `lbf` for pound force.

I've some many ways found this unit and struggled to convert it to
something sane like `N m` or `N cm`.

Appreciate any comments as I've never written any Purescript before.